expander norm ok
