+++ Makefile
@@ -6,7 +6,7 @@
 #    By: jareste- <jareste-@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/31 09:20:31 by jareste-          #+#    #+#              #
-#    Updated: 2023/08/26 10:44:11 by jareste-         ###   ########.fr        #
+#    Updated: 2023/08/26 21:29:00 by jareste-         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -35,6 +35,7 @@ EXPANDER_PATH = ./SRC/expander
 
 SRCS += $(addprefix $(EXPANDER_PATH)/, expander.c \
 		expand_utils.c \
+		expand_utils2.c \
 		expand_dots.c) 
 
 BUILTINS_PATH = ./SRC/builtins
